topics_ingestor/requirements.txt
spacy
cassandra-driver
python-dateutil
gevent